Make build_key 64-bit. (#31365)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/31362

### Problem description
We should capture all compilation flags into build key or compile hash.
Making build_key 64 bits makes it easier because a [similar attempt ran
into entropy loss
problem](https://github.com/tenstorrent/tt-metal/pull/29986#discussion_r2434300229)
with 32-bit build_key.

### What's changed
Made build_key 64-bit.  There should be no functional changes.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/18890126263/job/53919825414)
CI passes
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/18890127611)
CI with demo tests passes (if applicable)

+++ tt_metal/impl/debug/inspector/inspector.cpp
@@ -91,7 +91,7 @@ void Inspector::program_destroyed(const detail::ProgramImpl* program) noexcept {
 }
 
 void Inspector::program_compile_started(
-    const detail::ProgramImpl* program, const IDevice* device, uint32_t build_key) noexcept {
+    const detail::ProgramImpl* program, const IDevice* device, uint64_t build_key) noexcept {
     if (!is_enabled()) {
         return;
     }
@@ -107,7 +107,7 @@ void Inspector::program_compile_started(
 }
 
 void Inspector::program_compile_already_exists(
-    const detail::ProgramImpl* program, const IDevice* device, uint32_t build_key) noexcept {
+    const detail::ProgramImpl* program, const IDevice* device, uint64_t build_key) noexcept {
     if (!is_enabled()) {
         return;
     }
@@ -147,7 +147,7 @@ void Inspector::program_kernel_compile_finished(
 }
 
 void Inspector::program_compile_finished(
-    const detail::ProgramImpl* program, const IDevice* device, uint32_t build_key) noexcept {
+    const detail::ProgramImpl* program, const IDevice* device, uint64_t build_key) noexcept {
     if (!is_enabled()) {
         return;
     }
